refactor!: remove deprecated GeneratedVaadinContextMenu
